Created a workaround for next/prev bug
Adding left padding to the first element (first image) caused the
carousel to only show `slides.length - 2` images.

The `maxOffset` value was not large enough to show the two last images.
This caused the carousel to skip an image when going back after not
being able to set the index to the index of the last element.